backend/requirements.txt
fastapi
uvicorn
twilio
requests
python-dotenv
